Extract festafacility from gui code
move to core project as a structure/class
